fix(scout-for-lol): reconcile prematch changes with main
Why: Restacking PR #2301 onto main exposed a loading-screen conflict and a stale generated test database.\n\nWhat: Extract the prematch loading-screen renderer into its own module and regenerate the SQLite test template from the merged Prisma schema.\n\nVerification: bunx tsc --noEmit -p packages/scout-for-lol/packages/backend/tsconfig.json; targeted Bun integration tests (18 passed, 6 skipped); ESLint targeted files; bunx lefthook run pre-commit.
