+++ packages/scout-for-lol/packages/backend/src/league/tasks/prematch/prematch-notification.ts
@@ -1,4 +1,4 @@
-import { AttachmentBuilder, EmbedBuilder } from "discord.js";
+import { EmbedBuilder } from "discord.js";
 import type {
   RawCurrentGameInfo,
   PlayerConfigEntry,
@@ -6,6 +6,7 @@ import type {
   DiscordGuildId,
   QueueType,
 } from "@scout-for-lol/data/index.ts";
+import type { AttachmentBuilder, MessageCreateOptions } from "discord.js";
 import {
   resolveQueueTypeFromGame,
   queueTypeToDisplayString,
@@ -18,21 +19,6 @@ import { getChampionDisplayName } from "#src/utils/champion.ts";
 import { createLogger } from "#src/logger.ts";
 import { uniqueBy } from "remeda";
 import * as Sentry from "@sentry/bun";
-import {
-  RecoverableLoadingScreenDataError,
-  UnsupportedLoadingScreenQueueError,
-  buildLoadingScreenData,
-} from "#src/league/tasks/prematch/loading-screen-builder.ts";
-import {
-  loadingScreenToImage,
-  loadingScreenToSvg,
-} from "@scout-for-lol/report";
-import { savePrematchImageToS3, savePrematchSvgToS3 } from "#src/storage/s3.ts";
-import {
-  classicAssetResolutionFailuresTotal,
-  prematchLoadingScreenGeneratedTotal,
-  prematchLoadingScreenDurationSeconds,
-} from "#src/metrics/index.ts";
 import { recordCoreOutputsDelivered } from "#src/analytics/guild-lifecycle.ts";
 import { recordPoolMessageRefs } from "#src/betting/pool-open.ts";
 import { refreshBucksMessages } from "#src/betting/message-refresh.ts";
@@ -42,8 +28,8 @@ import {
   type BucksPrematchAttachment,
 } from "#src/betting/prematch-hook.ts";
 import { startParlayGeneration } from "#src/betting/parlay-generate.ts";
-import type { MessageCreateOptions } from "discord.js";
 import type { LoadingScreenData } from "@scout-for-lol/data/index.ts";
+import { renderPrematchLoadingScreen } from "#src/league/tasks/prematch/prematch-loading-screen.ts";
 
 const logger = createLogger("prematch-notification");
 
@@ -174,139 +160,6 @@ function buildPrematchPayload(input: {
       : {}),
     embeds: [input.fallbackEmbed()],
     components,
-  };
-}
-
-/**
- * Render the loading-screen image, or fall back to text.
- *
- * The image path carries its own metrics, Sentry fingerprinting, and
- * fire-and-forget S3 writes. Keeping it separate leaves notification delivery
- * responsible only for sending and recording messages.
- */
-async function renderPrematchLoadingScreen(input: {
-  gameInfo: RawCurrentGameInfo;
-  trackedPlayers: PlayerConfigEntry[];
-  queueType: QueueType | undefined;
-  gameId: string;
-  aliases: string[];
-}): Promise<{
-  attachment: AttachmentBuilder | undefined;
-  embed: EmbedBuilder | undefined;
-  data: LoadingScreenData | undefined;
-}> {
-  const { gameInfo, trackedPlayers, queueType, gameId, aliases } = input;
-  let loadingScreenAttachment: AttachmentBuilder | undefined;
-  let loadingScreenEmbed: EmbedBuilder | undefined;
-  let loadingScreenData: LoadingScreenData | undefined;
-  try {
-    const startTime = Date.now();
-    const firstPlayer = trackedPlayers[0];
-    if (firstPlayer === undefined) {
-      throw new Error(`No tracked players provided for game ${gameId}`);
-    }
-    const region = firstPlayer.league.leagueAccount.region;
-    const trackedPuuidSet = new Set(
-      trackedPlayers.map((p) => p.league.leagueAccount.puuid),
-    );
-
-    loadingScreenData = await buildLoadingScreenData(
-      gameInfo,
-      trackedPuuidSet,
-      region,
-    );
-    const [image, svg] = await Promise.all([
-      loadingScreenToImage(loadingScreenData),
-      loadingScreenToSvg(loadingScreenData),
-    ]);
-
-    const attachmentName = `loading-screen-${gameId}.png`;
-    loadingScreenAttachment = new AttachmentBuilder(Buffer.from(image)).setName(
-      attachmentName,
-    );
-    loadingScreenEmbed = new EmbedBuilder({
-      image: { url: `attachment://${attachmentName}` },
-    });
-
-    const duration = (Date.now() - startTime) / 1000;
-    prematchLoadingScreenDurationSeconds.observe(duration);
-    prematchLoadingScreenGeneratedTotal.inc({
-      queue_type: queueType ?? "unknown",
-      status: "success",
-    });
-    logger.info(
-      `[sendPrematchNotification] 🖼️ Loading screen generated in ${duration.toFixed(1)}s for game ${gameId}`,
-    );
-
-    void (async () => {
-      try {
-        await Promise.all([
-          savePrematchImageToS3(
-            gameInfo.gameId,
-            image,
-            queueType ?? "unknown",
-            aliases,
-          ),
-          savePrematchSvgToS3(
-            gameInfo.gameId,
-            svg,
-            queueType ?? "unknown",
-            aliases,
-          ),
-        ]);
-      } catch (s3Error) {
-        logger.error(
-          `[sendPrematchNotification] Failed to save prematch assets to S3:`,
-          s3Error,
-        );
-      }
-    })();
-  } catch (error) {
-    recordClassicLoadingScreenFailure(loadingScreenData, error);
-    const isRecoverable = error instanceof RecoverableLoadingScreenDataError;
-    prematchLoadingScreenGeneratedTotal.inc({
-      queue_type: queueType ?? "unknown",
-      status: isRecoverable ? "fallback" : "error",
-    });
-    logger.error(
-      `[sendPrematchNotification] ❌ Failed to generate loading screen for game ${gameId}:`,
-      error,
-    );
-    if (!isRecoverable) {
-      const context =
-        error instanceof UnsupportedLoadingScreenQueueError
-          ? {
-              fingerprint: [
-                "prematch-unsupported-queue",
-                gameInfo.gameQueueConfigId.toString(),
-                gameInfo.gameMode,
-                gameInfo.mapId.toString(),
-              ],
-              tags: {
-                source: "prematch-loading-screen",
-                gameId,
-                gameQueueConfigId: gameInfo.gameQueueConfigId.toString(),
-                mapId: gameInfo.mapId.toString(),
-                gameMode: gameInfo.gameMode,
-              },
-            }
-          : {
-              tags: {
-                source: "prematch-loading-screen",
-                gameId,
-                gameQueueConfigId: gameInfo.gameQueueConfigId.toString(),
-                mapId: gameInfo.mapId.toString(),
-                gameMode: gameInfo.gameMode,
-              },
-            };
-      Sentry.captureException(error, context);
-    }
-  }
-
-  return {
-    attachment: loadingScreenAttachment,
-    embed: loadingScreenEmbed,
-    data: loadingScreenData,
   };
 }
 
@@ -532,24 +385,4 @@ export async function sendPrematchNotification(
     `[sendPrematchNotification] ✅ Notifications sent for game ${gameId}`,
   );
   return delivery.sentMessageIds;
-}
-
-function recordClassicLoadingScreenFailure(
-  loadingScreenData: LoadingScreenData | undefined,
-  error: unknown,
-): void {
-  if (loadingScreenData?.layout !== "classic") return;
-  classicAssetResolutionFailuresTotal.inc({
-    phase: "prematch",
-    reason: "asset",
-  });
-  logger.error(
-    "Classic prematch loading-screen asset rendering failed",
-    error,
-    {
-      championIds: loadingScreenData.participants.map(
-        (participant) => participant.championId,
-      ),
-    },
-  );
 }
